DSP-10 extract csu regions
