+++ index.html
@@ -12,7 +12,7 @@ Any questions can be sent to joelnwalkley@gmail.com
     <title>DJJ Continuum Provider Map</title>
     <script src="mainmapdata.js"></script>
     <script src="statemap.js"></script>
-    <script src="./dist/main.js"></script>"
+    <script type="module" src="./src/initMap.js"></script>"
     <link rel="stylesheet" type="text/css" href="pdstyle.css">
     <script>
         (function (i, s, o, g, r, a, m) {
